Fix after reviewing (#38); Modify TokenInfo data models

+++ server/cache/cache.go
@@ -59,8 +59,8 @@ type Client interface {
 	IsRequestToCoinMarket(ctx context.Context) bool
 	TokenInfo(ctx context.Context) (*types.TokenInfo, error)
 	UpdateTokenInfo(ctx context.Context, tokenInfo *types.TokenInfo) error
-	CirculatingSupply(ctx context.Context) (int, error)
-	UpdateCirculatingSupply(ctx context.Context, cirSup int) error
+	CirculatingSupply(ctx context.Context) (int64, error)
+	UpdateCirculatingSupply(ctx context.Context, cirSup int64) error
 }
 
 func New(cfg Config) (Client, error) {
